changed cluster amplitude from uint16_t to uint8_t

+++ CalibTracker/SiStripLorentzAngle/src/SiStripLAProfileBooker.cc
@@ -338,11 +338,11 @@ void SiStripLAProfileBooker::analyze(const edm::Event& e, const edm::EventSetup&
             const GlobalPoint monogposition = (monodet->surface()).toGlobal(monoposition);
 	    ClSize = (monocluster->amplitudes()).size();
 	    
-	    const std::vector<uint16_t> amplitudes = monocluster->amplitudes();
+	    const std::vector<uint8_t> amplitudes = monocluster->amplitudes();
 	    barycenter = monocluster->barycenter()- 0.5; 
 	    uint16_t FirstStrip = monocluster->firstStrip();
-	    std::vector<uint16_t>::const_iterator idigi;
-	    std::vector<uint16_t>::const_iterator begin=amplitudes.begin();
+	    std::vector<uint8_t>::const_iterator idigi;
+	    std::vector<uint8_t>::const_iterator begin=amplitudes.begin();
 	    nstrip=0;
 	    for(idigi=begin; idigi!=amplitudes.end(); idigi++){
 	    sumx+=pow(((FirstStrip+idigi-begin)-barycenter),2)*(*idigi);
@@ -430,11 +430,11 @@ void SiStripLAProfileBooker::analyze(const edm::Event& e, const edm::EventSetup&
 	    
 	    ClSize = (stereocluster->amplitudes()).size();
 	    
-	    const std::vector<uint16_t> amplitudes = stereocluster->amplitudes();
+	    const std::vector<uint8_t> amplitudes = stereocluster->amplitudes();
 	    barycenter = stereocluster->barycenter()- 0.5; 
 	    uint16_t FirstStrip = stereocluster->firstStrip();
-	    std::vector<uint16_t>::const_iterator idigi;
-	    std::vector<uint16_t>::const_iterator begin=amplitudes.begin();
+	    std::vector<uint8_t>::const_iterator idigi;
+	    std::vector<uint8_t>::const_iterator begin=amplitudes.begin();
 	    nstrip=0;
 	    for(idigi=begin; idigi!=amplitudes.end(); idigi++){
 	    sumx+=pow(((FirstStrip+idigi-begin)-barycenter),2)*(*idigi);
@@ -527,11 +527,11 @@ void SiStripLAProfileBooker::analyze(const edm::Event& e, const edm::EventSetup&
 	    
 	    ClSize = (cluster->amplitudes()).size();
 	    
-	    const std::vector<uint16_t> amplitudes = cluster->amplitudes();
+	    const std::vector<uint8_t> amplitudes = cluster->amplitudes();
 	    barycenter = cluster->barycenter()- 0.5; 
 	    uint16_t FirstStrip = cluster->firstStrip();
-	    std::vector<uint16_t>::const_iterator idigi;
-	    std::vector<uint16_t>::const_iterator begin=amplitudes.begin();
+	    std::vector<uint8_t>::const_iterator idigi;
+	    std::vector<uint8_t>::const_iterator begin=amplitudes.begin();
 	    nstrip=0;
 	    for(idigi=begin; idigi!=amplitudes.end(); idigi++){
 	    sumx+=pow(((FirstStrip+idigi-begin)-barycenter),2)*(*idigi);
